# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile with experience history
      description: Returns a user profile with personal and professional experience arrays
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3
            pattern: "^[a-z0-9_]+$"
      responses:
        "200":
          description: User profile with experiences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileFull"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/check-username/{username}:
    get:
      operationId: checkUsername
      tags: [users]
      summary: Check username availability
      description: Returns whether a username is available
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3
            pattern: "^[a-z0-9_]+$"
      responses:
        "200":
          description: Username availability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameAvailability"
  /users/{username}/upvote:
    post:
      operationId: toggleUpvote
      tags: [users]
      summary: Toggle upvote on a user
      description: Toggles an upvote on the target user. One upvote per user pair.
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
            minLength: 3
      responses:
        "200":
          description: Upvote toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpvoteResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        city:
          type: string
        country:
          type: string
        interests:
          type: array
          items:
            type: string
        role:
          type: string
        bio:
          type: string
          nullable: true
        xp:
          type: integer
        upvote_count:
          type: integer
        created_at:
          type: string
      required:
        - id
        - name
        - username
        - city
        - country
        - interests
        - role
        - xp
        - upvote_count
        - created_at
    UserProfileFull:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        city:
          type: string
        country:
          type: string
        interests:
          type: array
          items:
            type: string
        role:
          type: string
        bio:
          type: string
          nullable: true
        xp:
          type: integer
        upvote_count:
          type: integer
        has_upvoted:
          type: boolean
        created_at:
          type: string
        personal_experiences:
          type: array
          items:
            $ref: "#/components/schemas/UserExperienceCard"
        professional_experiences:
          type: array
          items:
            $ref: "#/components/schemas/UserExperienceCard"
      required:
        - id
        - name
        - username
        - city
        - country
        - interests
        - role
        - xp
        - upvote_count
        - has_upvoted
        - created_at
        - personal_experiences
        - professional_experiences
    UserExperienceCard:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [personal, professional]
        category:
          type: string
        role:
          type: string
          enum: [hosted, joined]
        status:
          type: string
          enum: [joined, completed]
        xp_earned:
          type: integer
        date:
          type: string
        city:
          type: string
      required:
        - id
        - title
        - type
        - category
        - role
        - status
        - xp_earned
        - date
        - city
    UsernameAvailability:
      type: object
      properties:
        available:
          type: boolean
        username:
          type: string
      required:
        - available
        - username
    UpvoteResponse:
      type: object
      properties:
        upvoted:
          type: boolean
        upvote_count:
          type: integer
      required:
        - upvoted
        - upvote_count
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message