# lib/api-zod/src/generated/types/userExperienceCard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserExperienceCardRole } from "./userExperienceCardRole";
import type { UserExperienceCardStatus } from "./userExperienceCardStatus";
import type { UserExperienceCardType } from "./userExperienceCardType";

export interface UserExperienceCard {
  id: number;
  title: string;
  type: UserExperienceCardType;
  category: string;
  role: UserExperienceCardRole;
  status: UserExperienceCardStatus;
  xp_earned: number;
  date: string;
  city: string;
}