# lib/api-zod/src/generated/types/userProfileFull.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserExperienceCard } from "./userExperienceCard";

export interface UserProfileFull {
  id: number;
  name: string;
  username: string;
  city: string;
  country: string;
  interests: string[];
  role: string;
  bio?: string | null;
  xp: number;
  upvote_count: number;
  has_upvoted: boolean;
  created_at: string;
  personal_experiences: UserExperienceCard[];
  professional_experiences: UserExperienceCard[];
}